Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double f;
-    scanf("%lf",&f);
-    printf("You entered: %.4lf",f);
+    float f;
+    scanf("%f",&f);
+    printf ("You entered: %f",f);
     return 0;
 }
